refactor mutation hook
