+++ package.json
@@ -27,6 +27,7 @@
     "webpack-dev-server": "^3.8.0"
   },
   "dependencies": {
+    "@apollo/react-hooks": "^3.0.1",
     "apollo-boost": "^0.4.4",
     "axios": "^0.19.0",
     "graphql": "^14.4.2",
