+++ src/components/PhotoCard/index.js
@@ -1,15 +1,11 @@
 import React, { Fragment } from 'react'
-import { useLocalStorage } from '../../hooks/useLocalStorage'
 import { useNearScreen } from '../../hooks/useNearScreen'
-import { ToggleLikeMutation } from '../../container/ToggleLikeMutation'
-import { FavButton } from '../FavButton'
+import { ToggleLike } from '../../container/ToggleLikeMutation'
 import { ImgWrapper, Img, Article } from './styles'
 
 const DEFAULT_IMAGE = 'https://images.unsplash.com/photo-1518791841217-8f162f1e1131?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=60'
 
 export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
-  const key = `like-${id}`
-  const [liked, setLiked] = useLocalStorage(key, false)
   const [show, ref] = useNearScreen()
 
   return (
@@ -21,17 +17,7 @@ export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
               <Img src={src} />
             </ImgWrapper>
           </a>
-          <ToggleLikeMutation>
-            {
-              (toggleLike) => {
-                const handleFavClick = () => {
-                  !liked && toggleLike({ variables: { input: { id } } })
-                  setLiked(!liked)
-                }
-                return <FavButton liked={liked} likes={likes} onClick={handleFavClick} />
-              }
-            }
-          </ToggleLikeMutation>
+          <ToggleLike id={id} likes={likes} />
         </Fragment>
       }
     </Article>
