+++ src/container/PhotoCardWithQuery.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { PhotoCard } from '../components/PhotoCard'
 import { gql } from 'apollo-boost'
-import { Query } from 'react-apollo'
+import { useQuery } from '@apollo/react-hooks'
 
 const GET_SINGLE_PHOTO = gql`
 query getSinglePhoto($id:ID!) {
@@ -15,17 +15,12 @@ query getSinglePhoto($id:ID!) {
     }
   }
 `
-
-const renderProp = ({ loading, error, data }) => {
+export const PhotoCardWithQuery = ({ id }) => {
+  const { loading, error, data } = useQuery(GET_SINGLE_PHOTO, {
+    variables: { id }
+  })
   if (loading) return <p>Loading...</p>
   if (error) return <p>Error!</p>
-
   const { photo = {} } = data
   return <PhotoCard {...photo} />
 }
-
-export const PhotoCardWithQuery = ({ id }) => (
-  <Query query={GET_SINGLE_PHOTO} variables={{ id }} >
-    {renderProp}
-  </Query>
-)
